TV Detection + Landscape Enable for Setup Screen
